thermal: mediatek: add missing (u32) cast for GENMASK
When building with -Werror, the following build error is reported:
src/drivers/thermal/mtk_thermal.c:530:29: error: implicit conversion from 'unsigned long' to 'u32' (aka 'unsigned int') changes value from 18446744072904245247 to 3489660927 [-Werror,-Wconstant-conversion]
        .apmixed_buffer_ctl_mask = ~GENMASK(29, 28),
                                   ^~~~~~~~~~~~~~~~

Fix it by adding the missing cast.
Note that the upstream version [1] already has this cast.

[1] https://lore.kernel.org/linux-arm-kernel/20211019093404.1913357-5-msp@baylibre.com/
Fixes: dd466a555b80 ("thermal: mediatek: add MT8365 thermal driver support")

+++ drivers/thermal/mtk_thermal.c
@@ -527,7 +527,7 @@ static const struct mtk_thermal_data mt8365_thermal_data = {
 	.sensor_mux_values = mt8365_mux_values,
 	.version = MTK_THERMAL_V1,
 	.apmixed_buffer_ctl_reg = APMIXED_SYS_TS_CON0,
-	.apmixed_buffer_ctl_mask = ~GENMASK(29, 28),
+	.apmixed_buffer_ctl_mask = ~(u32)GENMASK(29, 28),
 	.apmixed_buffer_ctl_set = 0,
 };
